+++ app/models/ahoy_captain/filter_parser.rb
@@ -26,24 +26,15 @@ module AhoyCaptain
 
     def parse
       @filter_params.each do |key, values|
-        item = Item.new
-
-        item.values = Array(values)
-        item.predicate = Ransack::Predicate.detect_and_strip_from_string!(key.dup)
-        item.column = key.delete_suffix("_#{item.predicate}")
-        modal_name = AhoyCaptain.config.filters.detect { |_, filters| filters.include?(item.column) }
-        if modal_name
-          item.modal = modal_name[1].modal_name
-        end
-
-        label = if item.column == "goal"
-          AhoyCaptain.config.goals[values].title
+        if key == "properties_json_cont"
+          json = JSON.parse(@params[:q][key])
+          json.each do |k,v|
+            @items[key] = build_json_item("properties_json_cont", k, v)
+          end
+          next
         else
-          item.values.to_sentence(last_word_connector: " or ")
+          item = build_item(key, values)
         end
-        item.label = label
-        item.description = "#{item.column.titleize} #{::AhoyCaptain::PredicateLabel[item.predicate]} #{label}"
-        item.url = build_url(key, values)
         @items[key] = item
       end
 
@@ -51,6 +42,50 @@ module AhoyCaptain
     end
 
     private
+
+    def build_item(key, values)
+      item = Item.new
+      item.values = Array(values)
+      item.predicate = Ransack::Predicate.detect_and_strip_from_string!(key.dup)
+      item.column = key.delete_suffix("_#{item.predicate}")
+      modal_name = AhoyCaptain.config.filters.detect { |_, filters| filters.include?(item.column) }
+      if modal_name
+        item.modal = modal_name[1].modal_name
+      end
+
+      label = if item.column == "goal"
+                AhoyCaptain.config.goals[values].title
+              else
+                item.values.to_sentence(last_word_connector: " or ")
+              end
+      item.label = label
+      item.description = "#{item.column.titleize} #{::AhoyCaptain::PredicateLabel[item.predicate]} #{label}"
+      item.url = build_url(key, values)
+      item
+    end
+
+    def build_json_item(root, key, values)
+      item = Item.new
+      item.values = Array(values)
+      item.predicate = Ransack::Predicate.detect_and_strip_from_string!(key.dup)
+      item.column = "Property"
+
+      item.label = "URL #{values}"
+      search_params = @request.query_parameters.deep_dup
+      search_params["q"][root] = JSON.parse(search_params["q"][root])
+      if search_params["q"][root][key].is_a?(Array)
+        search_params["q"][root][key].delete(values)
+      else
+        search_params["q"][root].delete(key)
+      end
+      if search_params["q"][root].empty?
+        search_params["q"].delete(root)
+      else
+        search_params["q"][root] = search_params["q"][root].to_json
+      end
+      item.url = @request.path + "?" + search_params.to_query
+      item
+    end
 
     def build_url(name, values)
       search_params = @request.query_parameters.deep_dup